requirements.txt
greenlet==2.0.2
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dotenv==1.0.0
shiboken6==6.5.1.1
SQLAlchemy==1.4.47
